Avoid circular imports
by moving {App,Job,Tool} Info classes to dependencies.py
