+++ lib/galaxy/jobs/rule_helper.py
@@ -9,6 +9,7 @@ from galaxy import (
     model,
     util
 )
+from galaxy.tools.deps.dependencies import ToolInfo
 
 log = logging.getLogger(__name__)
 
@@ -42,9 +43,7 @@ class RuleHelper(object):
         else:
             # Have a Job object.
             tool = self.app.toolbox.get_tool(job_or_tool.tool_id, tool_version=job_or_tool.tool_version)
-        # Can't import at top because circular import between galaxy.tools and galaxy.jobs.
-        import galaxy.tools.deps.containers
-        tool_info = galaxy.tools.deps.containers.ToolInfo(tool.containers, tool.requirements, tool.requires_galaxy_python_environment, tool.docker_env_pass_through)
+        tool_info = ToolInfo(tool.containers, tool.requirements, tool.requires_galaxy_python_environment, tool.docker_env_pass_through)
         container_description = self.app.container_finder.find_best_container_description(["docker"], tool_info)
         return container_description is not None
