+++ lib/galaxy/jobs/runners/__init__.py
@@ -25,6 +25,10 @@ from galaxy.jobs.runners.util.job_script import (
     job_script,
     write_script
 )
+from galaxy.tools.deps.dependencies import (
+    JobInfo,
+    ToolInfo
+)
 from galaxy.util import (
     DATABASE_MAX_STRING_SIZE,
     ExecutionTimer,
@@ -394,9 +398,8 @@ class BaseJobRunner(object):
             compute_tmp_directory = job_wrapper.tmp_directory()
 
         tool = job_wrapper.tool
-        from galaxy.tools.deps import containers
-        tool_info = containers.ToolInfo(tool.containers, tool.requirements, tool.requires_galaxy_python_environment, tool.docker_env_pass_through)
-        job_info = containers.JobInfo(
+        tool_info = ToolInfo(tool.containers, tool.requirements, tool.requires_galaxy_python_environment, tool.docker_env_pass_through)
+        job_info = JobInfo(
             compute_working_directory,
             compute_tool_directory,
             compute_job_directory,
