+++ lib/galaxy/tools/deps/containers.py
@@ -266,58 +266,6 @@ class ContainerRegistry(object):
         return None
 
 
-class AppInfo(object):
-
-    def __init__(
-        self,
-        galaxy_root_dir=None,
-        default_file_path=None,
-        outputs_to_working_directory=False,
-        container_image_cache_path=None,
-        library_import_dir=None,
-        enable_beta_mulled_containers=False,
-        containers_resolvers_config_file=None,
-        involucro_path=None,
-        involucro_auto_init=True,
-    ):
-        self.galaxy_root_dir = galaxy_root_dir
-        self.default_file_path = default_file_path
-        # TODO: Vary default value for docker_volumes based on this...
-        self.outputs_to_working_directory = outputs_to_working_directory
-        self.container_image_cache_path = container_image_cache_path
-        self.library_import_dir = library_import_dir
-        self.enable_beta_mulled_containers = enable_beta_mulled_containers
-        self.containers_resolvers_config_file = containers_resolvers_config_file
-        self.involucro_path = involucro_path
-        self.involucro_auto_init = involucro_auto_init
-
-
-class ToolInfo(object):
-    # TODO: Introduce tool XML syntax to annotate the optional environment
-    # variables they can consume (e.g. JVM options, license keys, etc..)
-    # and add these to env_path_through
-
-    def __init__(self, container_descriptions=[], requirements=[], requires_galaxy_python_environment=False, env_pass_through=["GALAXY_SLOTS"]):
-        self.container_descriptions = container_descriptions
-        self.requirements = requirements
-        self.requires_galaxy_python_environment = requires_galaxy_python_environment
-        self.env_pass_through = env_pass_through
-
-
-class JobInfo(object):
-
-    def __init__(
-        self, working_directory, tool_directory, job_directory, tmp_directory, job_directory_type
-    ):
-        self.working_directory = working_directory
-        self.job_directory = job_directory
-        # Tool files may be remote staged - so this is unintuitively a property
-        # of the job not of the tool.
-        self.tool_directory = tool_directory
-        self.tmp_directory = tmp_directory
-        self.job_directory_type = job_directory_type  # "galaxy" or "pulsar"
-
-
 @six.add_metaclass(ABCMeta)
 class Container(object):
